+++ CHANGELOG.md
@@ -85,6 +85,7 @@
 - [10502](https://github.com/vegaprotocol/vega/issues/10502) - Add `underlyingIndexPrice` to perpetual data.
 - [10523](https://github.com/vegaprotocol/vega/issues/10523) - Fix repeated games statistics for multiple recurring transfers.
 - [10527](https://github.com/vegaprotocol/vega/issues/10527) - Add support for `byte32` type in market proposal oracle definition.
+- [10563](https://github.com/vegaprotocol/vega/issues/10563) - Spam protection for create/update referral program.
 
 ### 🐛 Fixes